requirements.txt
Pillow
google-cloud-storage
google-cloud-documentai